+++ src/codeLensProvider.ts
@@ -33,12 +33,12 @@ export class MarkdownCodeLensProvider implements vscode.CodeLensProvider {
         codeLenses.push(new vscode.CodeLens(range, termCmd));
 
         // 3) Bind cell to result block
-        const bindCmd: vscode.Command = {
-          title: "Bind",
-          command: "drafty.bindBlock",
-          arguments: [range],
-        };
-        codeLenses.push(new vscode.CodeLens(range, bindCmd));
+        // const bindCmd: vscode.Command = {
+        //   title: "Bind",
+        //   command: "drafty.bindBlock",
+        //   arguments: [range],
+        // };
+        // codeLenses.push(new vscode.CodeLens(range, bindCmd));
       }
     }
     return codeLenses;