+++ src/extension.ts
@@ -22,14 +22,14 @@ export function activate(context: vscode.ExtensionContext) {
     "drafty.terminateBlock",
     (range: vscode.Range) => commands.terminateBlockHandler(context, range),
   );
-  const bindBlockCmd = vscode.commands.registerCommand(
-    "drafty.bindBlock",
-    (range: vscode.Range) => commands.bindBlockHandler(context, range),
-  );
+  // const bindBlockCmd = vscode.commands.registerCommand(
+  //   "drafty.bindBlock",
+  //   (range: vscode.Range) => commands.bindBlockHandler(context, range),
+  // );
 
 
   // Register commands and CodeLens provider
-  context.subscriptions.push(startSessionCmd, runBlockCmd, terminateBlockCmd, bindBlockCmd);
+  context.subscriptions.push(startSessionCmd, runBlockCmd, terminateBlockCmd);
   context.subscriptions.push(
     vscode.languages.registerCodeLensProvider(
       "markdown",